requirements.txt
flask
python-dotenv
openai
pydantic
requests